Fix disulfide dihedral generation for GROMOS force fields
The code to clean up proper dihedrals for force fields
that do not keep all dihedrals (realistically only GROMOS)
added bogus atom indices across the disulfide as assigned
by a hackblock. This is unlikely to have resulted in silent
errors since the atom indices were completely off, and it should
not have influenced other dihedrals or other force fields.

Since the routine in question was very difficult to understand,
I ended up rewriting and commenting the clean-up in detail.

Fixes #4188.