Move pull force application to separate function
This resolves the issue that force from transformation coordinates
were not applied to the atoms (but forces were reported correctly
in the ouput).
This also simplifies the AWH main call, as the force buffers and
masses no longer need to be passed.

The force constant for the transformation coordinate pull integration
test has been increased by a factor 10, so it actually fails when pull 
forces are not applied.

Fixes #4385