Fix CI by upgrading HeFFTe to 2.2.0
Heffte 2.1.0 didn't allow application to pass a CUDA stream on which computation can be performed and all computation was performed on default stream. This behavior has changed with 2.2.0.
This allows GROMACS to perform FFT on PME stream and avoids unnecessary synchronization.

Change container build scripts to use 2.2.0 Heffte branch.

This is a cherry-pick of !2300 (d332c03059cd7b441ff5e774432898417fe69779) from `master` to `release-2022`.

Refs #4150

Closes #4389