Make current slow growth behaviour consistent and communicate it better
In lambda dynamics/slow growth, lambda can be used to interpolate the lambda vector or to set its components directly if no lambda vector is provided by the user. In this latter case, lambda was allowed to be > 1, but it was silently kept within [0,1] if a lambda vector was specified. Moreover, setting the components of the lambda vector > 1 as a user produced an error.

Now, it is consistently ensured that lambda vector components are in [0,inf), but warnings are issued if a user provides settings that somehow result in lambda vector components being > 1. If soft-core potentials are used, lambda vector components for Coulomb and vdW are consistently enforced to be in [0,1] (errors are issued else). If lambda is used to interpolate a user-provided lambda vector, it is kept in [0,1]. If user input results in lambda leaving the above ranges during the simulation, lambda will be kept at the respective interval boundary, and warnings are issued from which simulation step on the lambda vector will not change anymore.

Fixes #3584.