Allow Gromacs to run on more than 64 threads by default
Avoids Gromacs dying on machines when starting more than 64 threads,
which is getting more common today - including our own tests.

This increases the default (hard) OpenMP thread limit from
64 to 128, since at least Roland has not seen any performance
drawbacks from that.

Second, we no longer attempt to start more threads than Gromacs
has been configured for. In principle it would have been cleanest
to limit gmx_omp_get_max_threads() to GMX_OPENMP_MAX_THREADS,
but unfortunately the first value is used to detect the total
number of threads for all ranks, while the latter is rather used
as a limit for the number of threads to start inside each rank.
All this threading code needs to cleaned up later, but to avoid
severely limiting MPI parallelism on these machines, for now we
instead need to apply the limit once we have adjusted for the
number of ranks in the higher-level routine.

Closes #4370.