SYCL: Allow using GPU FFT with DPC++/MKL
The underlying issues are mostly resolved in oneMKL 2022.0.1, but we
still support 2021.4.

Since the functionality depends on the version of dynamically linked
libraries, the GPU FFT is disabled by default to be on the safe side.
Setting `GMX_GPU_SYCL_USE_GPU_FFT` environment variable is required
to use this feature.

Also, oneMKL currently requires padding for the real 

Refs #4219, #3927.

The dependency on !2415 is not a hard requirement, just a way to harmonize two changes to the same bits of code.