Use correct number of atoms in GPU Update kernels
With multiple ranks, we were allocating buffers and launching update
kernels for the total number of atoms, which it should be the number of
local atoms.

This caused overallocation of GPU memory and resources, but should not
have led to incorrect results.

I consider it a genuine bug with a trivial fix, so targeting 2022 release. But since the impact is minimal, I'm open to other suggestions (2022.1 or 2023).

Fixes #4398