Disable GPU halo communication for small domains
The GPU-direct halo communication (enabled by default when
using theadMPI) results in fatal errors when domains
are to small. This resolves the issue by rather
disabling GPU halo communication for very small domains. 
For constant-volume simulations this is strictly based
on the number of pulses, while isotropic-coupling runs use
a margin of 10% to account for shrinking boxes, and general
pressure coupling have a margin of 100% to account for 
changing box shapes.

Properly fixes #4329