Revert "Use correct number of atoms in GPU Update kernels"
This reverts commit 4d9a6d110b614a299a5bab4120c87a04f9ac14c1 (MR !2523).

It was supposed to be a trivial fix, but things turned out to be more
complicated, and my testing was insufficient (#4401).

The fix is still needed, but the bug is not causing incorrect physics,
merely harmless sanitizer errors and slightly higher resource
consumption by kernels that are very fast regardless.

I suggest doing it in master or delaying till the patch release.
Right now is not the best time to fix such issues.

Refs #4398