Revert default usage of GPU-direct communication
Even with thread-MPI, the GPU-direct communcation feature
has not had the amount of testing we hoped, and we have
unfortunately seen a few bugs late in the beta phase. Given
that our testing is imperfect, there will likely be others,
so we should not have it enabled by default for all users
relying on thread-MPI until it has been more extensively
tested internally. It will still be possible to use
it by setting the environment variables.

Refs #3950, #4373, #4375, #2891, #2915.

Fixes #4400, #4329.