Added sharper warnings for Berendsen weak coupling
Since we now have Bussi coupling both for pressure and temperature
we should just tell new users to stay away from Berendsen. The
advanced ones will always be able to choose themselves.

Cherry pick from release-2022, so that the merge and change to
regression tests doesn't need to be at the same time.