Add symmetrization and centering to gmx potential
Since this analysis is frequently used for membranes,
it makes sense to support the same options as the
density analysis tool. This is admittedly a bit of
a hack, since we have just copied functionality
rather than properly refactored, but at some point
features also matter, and we can't postpone requests
forever due to refactoring.

Fixes #3579.