Add a check for k0 != 0 to avoid divide by zero error in distance restraint calculation.
This is a follow-up to MR !2456. I realized that I had overlooked the corner case k0 = 0, which would lead to a division by zero. This may also be related to bug #4414, where one failure occurs in NbLibListedForcesTests, which looks like it could include distance restraints.

I am very sorry if my oversight indeed caused this problem.