Use correct number of atoms in GPU Update kernels
With multiple ranks, we were allocating buffers and launching update
kernels for the total number of atoms, which should be the number of
local atoms.

This caused overallocation of GPU memory and resources, but should not
have led to incorrect results.

Early returns are added to prevent launching kernels with no work.

Also added assertions to the kernel launch wrappers so no zero atom case
should occur at that level.

Fixes #4398