Disentangle PME MPI buffer indices
The working buffer index was a member of SlabCommSetup, but used
an element of that class in PmeAtomComm indexed directly on slab
index instead of on offset. Now a separate vector is used for the
buffer indices so we have consistent indices.

Also renaming nodeidin PmeAtomComm to slabIndex for clarity.