Fixed two minor unit test errors discovered by Fedora testing
Fixed incorrect charmm27 patch atom type name, and
the code construct that made it result in a crash. This
would have generated errors later if pdb2gmx was
used with this patch, so it will not have lead to any
silent incorrect results.

The same change also fixed a pure internal crash in
nblib unit tests. This could never affect any usage of
nblib.

Fixed #4414.