Correctly check for FE lambda going over 1 during preprocessing
The checks for whether the value of lambda would go over 1 or under 0 during the simulation were giving a warning/an error when the delta-lambda and number of steps was "just right". This was due to truncation when attempting to compute the step where such lambda overflow would occur. Therefore, the step number is now calculated as a double and rounded before it is cast into an integer.

Adds unit tests for that particular check.

Fix #4442 and #4448