Remove `gmx_broadcast_world()` from `utility` module.
This simple wrapper is only used internally for initializing
CommandLineProgramContext. For this use case, its hard-coded
reference to MPI_COMM_WORLD is appropriate. For more general
use cases, the largest communicator that such a function
should use would be a `libraryWorldCommunicator`, such as
from `SimulationContext`.

We probably don't need a public utility function to do something as
simple as gmx_broadcast_world, but we would need more discussion and
progress with respect to ~external-comm and ~ResourceManagement2023
before reintroducing.

Ref #4457