Allow incomplete information when extracting clusters
Previously, all frames from an input trajectory needed to be defined to
be able to extract single frames using gmx extract-cluster.

This could lead to strange program behaviour when a user-created index
file with incomplete information was used (e.g. if not all frames
where defined, or fewer total frames).

Now the tool correctly skips frame numbers that are higher than the
highest one found in the index file.

Fixes #4420

Closes #4420