Fix ekindata checkpointing
With the md integrator, the checkpoint would not contain the
kinetic energy data when checkpoint occured at an energy calculation
step and energies were not calculated every step (which is the most
common case). This could lead to least significant bit differences
in the kinetic energy at t-dt/2, causing divergence, even with the
-reprod option.

Also removed several unused kinetic energy terms for the checkpoint
file. Likely more are unused, these have been documented in the code.

Fixes #4240