Fix unused-but-set-variable warning
Issued by Clang 15 pre-release when a variable is set but never used.

Targeting `release-2022-hwe` since it will become a problem with the next oneAPI release.
If there's a general consensus of supporting Clang 15 in the 2022 branch, I can retarget the MR.

Most changes are trivial removal of unused variables.

The only big change is in `mk_mshift`:

- Removed code hidden by macro that's never used.
- Add loop invariant assertion.