Use SYCL_CXX_FLAGS_EXTRA with hipSYCL
We use this CMake variable (introduced in !2252) to pass flags to oneAPI/IntelLLVM.

I see no reason to not use it for passing extra flags to hipSYCL as well.

Note: it is already possible to set `HIPSYCL_SYCLCC_EXTRA_ARGS`, and it will be updated correctly by CMake. But having a unified way of supplying flags is more convenient, IMHO.