Move topology information out of mtop_util
The header for mtop_util included the topology header for convenience
reasons that made refactoring of the atoms datastructures more difficult
due to the inclusion of this header in a number of places, including the
GPU code.

Moved the header out of it, made a dedicated header for enums used in
several places, moved helper functions in the GPU code to a dedicated
translation unit.

Refs #2833