Avoid inter-object-file dependencies
Compiling an object file should not depend on other object files.
Before this, when compiling an object file outside the modular
simulator after changing a header file that was also required by the
modular simulator module, the modular simulator had to be compiled
before the object file. Now it does not.