Add python script to run automated release builds
When run locally (on a developers machine), this script submits a new
release pipeline for the specified branch, with the script later
uploading the artefact files to the GROMACS project web servers, using
the same script in server mode.

Adds necessary triggers and pipeline specifications.

Also fix sign error in copyright checker script.