Remove inputrec from pressure-coupling code
Made PressureCouplingOptions to handle the fields of the inputrec that
pertain to pressure coupling. This means that pressure-coupling code
can be isolated from t_inputrec and is an early step to making a
pressure-coupling MdModule that can fill the options via the KVT
machinery. It also makes it easier to write unit tests for
pressure-coupling code.

Fixed a comment in coupling.cpp that used to have context that
got moved. Now it provides the information itself.