Bool clone that can have a view on std::vector
`vector<bool>` can't have a view as an `ArrayRef` on it, because it lacks `.data()` and the likes as it can be implemented as bitmask.

This MR introduces a type that behaves like a `bool` but does not trigger the template specialisation that `bool` does.

This change enables removal of legacy `bool *`