Properly release gmx_domdec_t
`gmx_domdec_t` already implements destructor, but we were not calling
it. We add a helper free function around it, to be able to easily deal
with forward declarations and raw pointers.

A better solution would be switching to `std::unique_ptr` / `gmx::unique_cptr`,
but it would require extensive changes in many places `cr->dd` is used,
in particular various `dd_*` collectives. We don't want that in "stable" branch.

Since `gmx_domdec_t` also contains GPU-related data through
`gpuHaloExchange`, we need to free it manually in `Mdrunner::mdrunner`
before the GPU context is destroyed.

Refs #4372, #4387