PME GPU: Don't copy uninitialized spline data
In `pme_gpu_copy_output_spread_atom_data`, we were copying whole
buffers, including padding, which contained uninitialized data.
This was detected by Initcheck (from NVIDIA's Compute Sanitizer).

To fix it, we only copy data for warps that contain atoms. Due to the
GPU data layout, we cannot just use nAtoms, but shall round it up to the
number of active warps.

Note: Copying uninitialized data is not an undefined behavior per
C++ standard, but is still a nasty thing to do.
The copied padding was not accessed on host, so it is not a real bug,
but it prevents us from using Initcheck (it does not support
suppressions).

Another solution would be to clear the buffers after allocation, but
that would add two more GPU operations, and could hide a future error
when the buffers are actually not getting initialized.

Refs #4453