Add post-merge CUDA Compute Sanitizer job
Run three of the tools (memcheck, racecheck, synccheck) on a
subset of test to ensure we get some coverage with reasonable run time.

Initcheck requires non-trivial fixes (!2652) and will be enabled later.

To keep the runtime low and to avoid complaints about no GPU being used, 
a new CTest label was introduced: `QuickGpuTest`. It is assigned to all unit 
tests (`gmx_add_unit_test`) which have `HARDWARE_DETECTION` argument, and can 
be manually added to the rest of the tests.

I ran the tests on my local machine with GTX1660SUPER (also doing graphics), 
and added this label to the Mdrun tests that complete in under 10 seconds when run with `memcheck`.

Example output (before #4393 was fixed): https://gitlab.com/gromacs/gromacs/-/jobs/2110775468

The job took 5 minutes on 1 GPU, so it should be OK to run it in post-merge.

Refs #4393, #4453