Add TestDevice::activate and make it context-centric
Instead of directly calling `setActiveDevice`, we now have
`TestDevice::activate`, which activates device's context,
which, in turn, calls `setActiveDevice`.

That will allow easier extension of this functionality later, which is
needed for host buffer pinning (otherwise, we'll have to do major
refactoring in order to pass `DeviceContext` to a lot of different
places).

And, besides, to me it seems like a nicer code.

Part of MR !2706.