Exclude top-level `*build*` directories from Docker build context.
If anyone is doing in-source builds, they might be doing multiple
configurations of in-source builds. In addition to `build` and `_build`,
match all top-level directories containing the string "build".

This can have a big effect on the start-up time of a Docker build
and can inappropriately expire build cache layers that should not
be expired.

!2754 updated `.gitignore`, but neglected `.dockerignore`