Fix mismatch and overflow in creating OpenMP thread-count env variable warnings
In commit 678937cb61a, we forgot to add a field to the array
describing the module that an environment variable matches. So anybody
using these environment variables for modules related to virtual sites
or constraints would have seen a confusing or garbage message.

Then in commit aa09f40a0689, we swapped the order of the strings used
in formatting a message, so the output would have been garbled as
well.