Support out-of-place GPU halo exchange
Removes limitation on in-place receives for GPU halo exchange. This is
only triggered in unnatural corner cases which are over-decomposed
with very small numbers of atoms per GPU (well below the threshold
known to be appropriate for performant multi-GPU scaling). Hence we opt
for a simple solution staging data via host memory, which allows
thread-MPI and lib-MPI paths to be unified. More optimal solutions
would lead to more complexity.

Refs #2891 #4329