Wait for PME coordinate padding clearing to complete on search steps
Previously, the clearing occured in the PME stream, but the subsequent
PME calculation proceeded in different streams if PME-PP pipelining
was enabled. In certain run configs, with libMPI the ordering of PME
coordinate receives could be such that intermittently the PME spread
kernel could commence before clearing, causing a crash.

Fixes #4482