Add functions to set context when pinning host memory
The context is not used in CUDA, but will be needed for SYCL host memory
pinning (and OpenCL, if we ever decide to implement it).

Ideally, we shall now pass the context to every `pmalloc`/`pfree` call.
But since that would require large code restructuring, we also add
functions to set the default context (per-thread) to use for host
pinning, which will be used unless a context is explicitly supplied.

Creating a new context or activating one automatically calls sets them
as a default.

Part of MR !2706.