Fix nightly SYCL build by switching to OpenCL.
FFT unit test started failing with LevelZero backend
after upgrading kernel and runtime on our CI machine.

By default, GROMACS uses OpenCL and does not use FFT, so it is not a
relevant combination for most users. Thus, we opt to simply not change
the backend used in that test from LevelZero to OpenCL.

We still run regression tests with LevelZero.

Added a general known issue.

Refs #4219.