Improve handling of nvcc flags when using gcc 7
The previous code detected valid nvcc flags as invalid when gcc was
the host compiler, probably leading to poor performance.

Now such flags are assumed to be valid, and the user provided with a
hint in a warning that if they have problems that they should use a
more recent gcc.

Closes #4478