GPU: Synchronize with the stream before destroying things
We are now synchronizing with relevant streams before destroying memory
buffers to make sure we have no scheduled memcpy's and memset's using
them.

A proper solution would be to not schedule such tasks (since their
results are not used anyway), but this might further complicate our
schedule by requiring to have a special treatment of the very last step.

From brief examination, only Local and Update streams are
actually affected, but it seems easier to sync with everything.

Refs #4519.