Use effective density for calculate the Verlet buffer
The worst approximation that can cause an underestimate of
the required Verlet buffet size for the pair list is homogeneous
atom density. Now an effective atom density is determined by an
atom count weighted average of the atom density in cell with the size
of the nonbonded cut-off. This uses the coordinates at mdrun startup.

Refs #4509