Set the code page to UTF8 for MSVC (and compatible) compilers
GROMACS assumes all is UTF8 (as it is on Unix) but didn't set it on Windows where it isn't the default.
It still assumes the console has UTF8 enabled (`chcp 65001`). Otherwise everything else (ie. file-io) works but console output is still incorrect.
It is possible to fix that by calling `SetConsoleOutputCP(CP_UTF8)` in `init`. But that changes the terminal even after the application has finished and resetting it to the previous code-page isn't possible if the application doesn't finish properly.