Update CUDA CI container to gcc 9
This prepares to update the required GCC version to 9

Updated the recipe for the container and used it in CI matrix.

Nvidia have changed their GPG signing strategy. Old containers can't
use apt-update without getting a GPG key from Nvidia. hpccm doesn't
support getting that key with the things that are already found in the
container.  So instead we update the minimum CUDA version tested to
11.0.3 because that container already has the right key in it.

Refs #4093
Refs #3708