Add test for keyvaluetree multi values
We need to make sure that we can create key-value tree entries based on
arbitrary number of user inputs for e.g. constant pH groups, pull groups
or AWH groups.

This adds a small test case to see that this indeed works with the kind
of input you could expect, but more advanced test cases are still
needed.