SYCL: fix HostAllocator test
For SYCL (and OpenCL), the pinned memory is accessible only from the
same context in which it is allocated. One of the host allocator tests,
`HostAllocatorTestCopyable/*.TransfersWithPinningWorkWithDevice`, was
using a queue for data transfer that was not bound to the allocation
context. This caused the test to fail with the OpenCL backend on intel-gpu02
(LevelZero backend and single-GPU machines were working fine).

Now, we pass the context explicitly to the helper test function, and
the queue is created using this context.

In CUDA, the context is ignored. In OpenCL, the context is also reused,
although it does not seem necessary, since we don't pin memory there.

Note: there are still regression tests on DG1, investigating,

Refs #4522