Fix the extlinks caption string.
As of Sphinx 4.0, the extlinks caption allows substitution of
exactly one `%s`. As of Sphinx 6.0, exactly one `%s` in the caption
string is _required_. As of Sphinx 5, failure to use the caption string
properly produces a warning.

This change means we need to update our minimum required Sphinx version
to Sphinx>=4.0.

Closes #4544

Closes #4544