Fix duplicate index group creation
The code that analysed and created the index groups contained a logic
error that caused the OTHER group to be analysed multiple times if
additional groups are present in the system that are not classified as
Protein or DNA/RNA. This would happen all the time when e.g. Ions had
been added to the system, as this group is not recognized by default.

Added explicit detection of the Ion group, and a check to make sure that
other group is only analysed once.

Fixes #4524