Use Matrix3x3 in BoxDeformation
Now that it has value semantics, permitted copy and move operations.

Also renamed the builder function to be consistent with the rest of
the code.

Moved several utility functions to boxmatrix.h, now that it has parts
that are used in three places.

Added several utility functions to matrix.h