Correctly set event consumption count with GPU-direct PP-PME comm
With GPU direct communications, the pmeForcesReady event is consumed once for each PP rank.
Now we correctly set this consumption count.

For GPU-aware libMPI, we have some extraneous host-side synchronizations which can
be avoided by synchronizing (and consuming the event) only once. But this solution
will not work for threadMPI and with future stream/event-aware libMPI.

Refs #3988, #4372