Improve SIMD detection and warnings
We now properly handle that 128-bit FMA4-based AVX that
is optimal for pre-Zen AND CPUs is not supported on more
modern CPUs. For parallel runs with non-matching hardware
we drop to plain AVX instead, and we have a special check
where we explain to users that this is a deprecated
SIMD type.

Fixes #4526.