Move math headers to include dir
This file movement paves the way for decoupling and eventual removal of libgromacs target. Once modules are uncoupled from libgromacs this chage will also make it easier to change math into an OBJECT target.

This change also includes several changes in linking of tests and modules that were needed for compilation.

Relates to #3288