Use persistent GpuEventSynchronizer's in PME pipelining
This avoids repeated calls to cudaEventCreate/cudaEventDestroy in the
PME rank for each PP rank on each step. With CUDA, the benefit is
small: approx. 5µs vs 9µs of API overhead per PP rank. For HIP, it is
expected to be bigger.

And overall, why make API calls when we can avoid doing it?

Refs #4572