Fix webpage error filtering removed in b065f3c7.
The goal of b065f3c7 was to build doxygen targets separately (before)
sphinx targets and to handle the errors separately. However, there does
not seem to be a way to build the `webpage` target without rebuilding
the `doxygen-all` target when GMX_VERSION_INFO is ON.

Instead, we use the `webpage-sphinx` target to collect sphinx errors,
and accept duplicate doxygen-all results for the `webpage` target.