Interface for frame conversion
Add IFrameConverter and module chaining method

Adds the IFrameConverter interface to perform operations on coordinates
contained in t_trxframe datastructures as prerequisites to performing
data analysis or writing coordinates to disk.

The first method implementing the interface is a dummy that allows
chaining methods together to be performed one after the other.

Refs #3555 

Implements part of #3556