Begin to reestablish public API documentation.
Set up a new branch of the Sphinx document tree. Provide an overview of
building client software.
(Import and update some of the legacy API docs.)

No new API documentation is exposed because we are still in the early
stages of figuring out an approval process for public API specification.
However,
* the build system documentation should be immediately useful to people
  building existing client software,
* the new documentation node provides a place in which to test ideas on
  API documentation, and
* this commit provides a baseline for content in support of
  #3777 and #4447

Ref #3402