Increase the default values of nsttcouple and nstpcouple
The default values of nsttcouple and nstpcouple have been increased
from 10 to 100 to improve performance of GPU and parallel runs.
The mechanism to lower the values to ensure accurate integration
is now documented. Now also v-rescale uses at least 5 integration
steps per tau, as all other first order coupling algorithms.

Also set nstcalcenergy to 100 instead of 10 when the user supplied
value is -1 (note that the default mdp value is 100).

Fixes #4565

Closes #4565