Move MDModules notifiers to high-level module
This partly resolves the circular dependency of math on utility.

The MDModules notifier functionality should live at a higher level. In
particular, the notification types should be in a utility module low
enough that e.g. grompp, mdrun, and the modules receiving the
notifications can all comfortably depend on them, and high enough that
they can depend on types in other utility modules. For example, RVec
and ArrayRef should be able to be used in notifications, but that
means the notification type must live above both math and utility
modules.

Refs #3288