Remove ISerializer overloads for rvec
Finishes breaking circular dependency of math and utility.

The serialization interface should live at a low level, but if it has
methods for rvec and ivec then it has to depend on math module. Then
there is no clear module at a level higher than math but lower than
everything else in which ISerializer should live.

Since it is not necessary to have convenience methods for these types,
they are removed. They were being used in the fileio module, which now
implements that functionality using free functions. 

Refs #3288