Bump minimum cmake version to 3.18
Cmake 3.18.4 was released on 6 October 2020. The default version of cmake
in Ubuntu 22.04 is 3.22.1, so there should not be a problem there.

This is necessary because cmake 3.17 introduces find_package(CUDAToolkit),
which is a major step forward in usability over find_package(CUDA) or
enable_language(CUDA). Further, the first time that support for automatically
adding compiler flags for cuda version was introduced in cmake 3.18.

This will also make packaging cuda code much easier, as
without this change, it would be necessary to handle the different
variables that are set by the old find_package(CUDA) and the newer
find_package(CUDAToolkit). This work would also be somewhat pointless
busywork since surely we would anyway bump to at least cmake 3.18 after
the release.

The actual changes to handling of cuda cmake code are left to a followup.

Relates to #3288, #4428 (closed)