Use classes for j-list handling
This prepares for future changes to the value of
c_nbnxnGpuNumClusterPerSupercluster and thus c_nbnxnGpuJGroupSize to
be more flexible in optimizing for a wider range of GPUs. This will be
easier to manage if the compile-time values come from templates rather
than being hard coded.

This commit begins to replace the concept of j4 clusters with packed j
clusters, since the number of elements packed per int may change in
future. The bulk of the renaming is deferred for a future change.

Refs #3847