Avoid delayed hardware de-initialization in tests
In our tests, `gmx_hw_info_t` was only destroyed at the very end of the
program as part of destroying the static `TestHardwareEnvironment`. Since
the order of destruction of static orders is unspecified, it was
sometimes destroyed after hipSYCL shut down the rest of the CUDA/HIP runtime.

With CUDA, it led to harmless `cudaErrorCudartUnloading` errors printed to
stderr. With HIP, however, it was sometimes causing segfaults in 
libamdhip64.so. Only in tests and only at the very end, but still not nice.

We now explicitly destroy `gmx_hw_info_t` after running all the tests.