SYCL: Remove deprecated is_host
It is deprecated in SYCL2020, and IntelLLVM now complains about it.

We had this check to avoid querying device properties for the `host`
device, which was throwing exceptions.
Since 9eab63bacd (!3126), we can handle exceptions just fine.

Moreover, since ecb223ff5 (!2653), we're only listing GPU devices, 
so the host device would never be passed to this function anyway.