SYCL: Fix harmless indexing error in spline calculation
Why was this not caught? Because it does not change the result.

I'll use the CUDA notation below.
Note that thread indexing is reversed, `threadIdx.x` is `itemIdx.get_local_id(ZZ)` etc.

In release-2022, only "order²" kernels were used, with `blockDim` = {4, 4, _atomsPerBlock_}.

Correct indexing: `threadIdx.y` * `blockDim.x` + `threadIdx.x` = [0; 4) * 4 + [0; 4) = [0; 16),
where "[a; b)" means any integer in this range, two ranges being independent.

Incorrect indexing: `threadIdx.y` * `grodDim.x` + `threadIdx.x` = [0; 4) * _N_ + [0; 4),
where _N_ is the padded number of atoms divided by _atomsPerBlock_.
The number of atoms is padded to 64, _atomsPerBlock_ is 8 or 16, so _N_>1.

Later, we only do anything if `threadLocalIdXY % DIM < 3` and `threadLocalIdXY / DIM < 1`,
so we only care about the case `threadIdx.y == 0`, so the error in the formula does not affect
anything.

For "order" kernels, `blockDim.y == 1`, so `threadIdx.y` is always 0, so again, does not matter.

Yet, incorrect indexing is bad to have in the code, so I'd rather change it in `release-2022`.