hipSYCL: improve enqueueWaitEvent
Previously, we were launching an empty kernel on the GPU each time we
used GpuEventSynchronizer::enqueueWaitEvent in hipSYCL.

Now, we are only enqueueing the barrier without any unnecessary GPU work.

Occasionally there are random test failures similar to #4290 on sm_86,
but that did not reproduce on other machines, and the divergences are
small (<100 ULP energy, with tolerance 50ULP).

Refs #4064