SYCL: Increase PME padding
With a 64-wide warp and Order data layout (used for large systems),
the block size was greater than data padding, causing out-of-bounds
access.

Does not affect SYCL in release-2022 and OpenCL since they were always
using Order² layout.

This is a quick fix to keep the HWE branch bug-free.
More flexible solution will be made for `main`.

Related to #4630.