Allow client software to inspect the CMake version used for GROMACS.
Recent issues and even CMake bugs have revealed that the CMake version
used for the GROMACS installation can have important effects on
installed binaries and client build systems. Since we can only fix bugs
or improve build system support for later installations, it may be
helpful to client software to be able to construct workarounds based on
inference of CMake behavior of existing installations of GROMACS 2023
and onwards.

See, for instance, #4563