Introduce a new CI job that runs the build and tests on Linux ARM64
This PR makes it possible to run the build and tests in more then one CI workflow.
It extracts the CI workflow steps in a .cmake script files.
Both the old CI job (x86_64, macOS and Windows) and the new one (aarch64, Linux) execute the same sequence of .cmake scripts.