Fix CUDA graph unit test crash when no GPU is available
The test did not check for the availability of a compatible device which
lead to a runtime segfault if a CUDA build was run on hardware without
GPU or with a GPU not available / compatible.
While it is not strictly needed to test the graph code on each device
given that we don't have a reason to expect different behavior, other
tests use this patterns in such cases so we iterate through all
compatible devices here as well.

Refs #4277