Add required sync for multi-GPU graphs
Previously, on graph launch where a graph was not used the previous
step, there was a sync to ensure GPU activities on the previous step
had finished, but only on the GPU local to the graph launch. This fix
adds the required sync to the corresponding activities on all other
GPUs used for the PP domain decomposition.

Refs #4277