Add ensemble temperature mdp options
This enables users to set a system wide ensemble temperature for
systems where not all atoms are coupled or have the same reference
temperature. It also gives safety by not allowing the user of
algorithms that require an ensemble temperature when this is
not available.

To make this work the group reference and system ensemble temperatures
are stored in gmx_ekindata_t. This removes the ugly modification of
ref_t in t_inputrec in mdrun. Now t_inputrec is really const in mdrun.

Refs #3854
Refs #4495

Closes #3854