Enable graph update rather than re-instantiation on single-GPU
Enables CUDA graph update on single-GPU, so long as PME tuning is not
active (which can change the FFT kernels chosen by the FFT library as
grid sizes change). Note that multi-GPU update is not yet supported
since the graphe resulting from multi-threaded has a non-deterministic
ordering of nodes requiring additional CUDA support which is not yet
available.

Refs #4277