Add fep exclusion check at pairlist generation
Implemented a fail-safe check for excluded perturbed pair beyond rlist
at pairlist generation time. The distance check in the free-energy
nonbonded kernel has been removed, as this is no longer needed.
Note that excluded pairs might drift to a different periodic image.
This is very unlikely and even if it does happen the error is very
small due to the form of the (PME) potential. This is documented
in the code.

Refs #4303
Refs #4321
Fixes #4461